# Compute Symbol.__hash__() once