Also note *why* we need `Symbol.__hash__()`


A while back, but I forget where, the justification for why we using Strings in function instead of  Symbols in the code was because `is` or equality testing was slow.  This simple change addresses that. 

I forget where the benchmarking data was posted for previous comparisons. I'd be interested though in what the change (if any) is.