@rocky 2022-12-04
> In any case, I am not sure if this makes the comparison faster: `is` compares memory locations, while `__equal__` compares the result of `get_sort_key()` method. In any case, go cal try to benchmark this.

In any event, let's please phase out the use of strings when symbols are meant.