@mmatera 2022-12-05
Here we go with some basic benchmarks. The point here is  ```hash(key)``` is called each time a dictionary is accessed using the key ```key```. 

Now, let's compare the time required for computing ```hash``` for  a ```str```, for a ```tuple``` of two ```str```, and for a reduced implementation implementation of the old and new ```Symbol``` class:

```
class SymbolOld:
    def __init__(self, name):
        self.name = name

    def __hash__(self):
        return hash(tuple(("Symbol", self.name)))


class SymbolNew:
    def __init__(self, name):
        self.name = name
        self.hashval = hash(tuple(("Symbol", self.name)))
        
    def __hash__(self):
        return self.hashval
```


| `str` | `tuple` | `SymbolOld` | `SymbolNew`  |
|----      | ----    |  ----        | -----       |
| 21.3  |  43.4   |   87.9      |   51.9         |

(times are in ns)

Times were estimated using timeit in Pyston 2.3.4, called from the command line:

```
python -m timeit   -s 'from atestclass import SymbolNew; a=SymbolNew("System`ASymbolName")'  'hash(a)'
```

Notice that the time required for calling ```hash``` in the cached version (i.e., which does not have into account the time required for actual evaluation of the hash of a tuple)  requires 51.9ns, against the 21ns required to hash a ```str```.