@rocky 2022-12-05
To be clear here . The difference we care about, or that is significant in overall execution between SymbolOld and SymbolNew  is operationally the difference between accessing a class variable (new) versus calling hash on a tuple of strings.

The object creation time isn't that important - it is a one-time thing.