@rocky 2022-12-07
I spent a little time trying benchmarking the difference. What I am seeing is not much difference between either old, new, or using a string  in pyston. In 3.10.8, I can see a little difference but not much to warrant perverting and complicating the code. And as soon as a conversion is used to get a string from a Symbol, the negligible performance "benefit" is lost. 

A big problem I notice with the code above is that it defines  `__init__()`  and that is wrong - it should use `__new__()`. 

```
$ pyenv local pyston-2.3.5
$ python s2-timing.py 
'func_old'  0.36 ms
'func_new'  0.34 ms
'func_str'  0.32 ms
$ python s2-timing.py 
'func_old'  0.36 ms
'func_new'  0.35 ms
'func_str'  0.32 ms
$pyenv local python-3.10.8
$ python s2-timing.py 
'func_old'  1.61 ms
'func_new'  1.44 ms
'func_str'  1.06 ms
$ python s2-timing.py 
'func_old'  1.82 ms
'func_new'  1.19 ms
'func_str'  1.03 ms
```

Here is the code for s2-timing.py

```python
from mathics.timing import timeit

class SymbolOld:
    def __new__(self, name):
        self.name = name

    def __hash__(self):
        return hash(tuple(("Symbol", self.name)))


class SymbolNew:
    def __new__(self, name):
        self.name = name
        self.hashval = hash(tuple(("Symbol", self.name)))

    def __hash__(self):
        return self.hashval

old = SymbolOld("old")
new = SymbolOld("new")

d = {old: 1, new: 2, "string": 3}

ITERATIONS=10000
@timeit
def func_old():
    x = 0
    for _ in range(ITERATIONS):
        x |= d[old]

@timeit
def func_new():
    x = 0
    for i in range(ITERATIONS):
        x |= d[new]

@timeit
def func_str():
    x = 0
    for i in range(ITERATIONS):
        x |= d["string"]

func_old()
func_new()
func_str()
```

Trying another way in pyston 2.3.5 (output is old, new, str): 

```$ python symbol-timing.py
0.0005529960617423058
0.0005363970994949341
0.0006319154053926468
$ python symbol-timing.py
0.0005638459697365761
0.0005182567983865738
0.0005195159465074539
$ python symbol-timing.py
0.0005540568381547928
0.0005348557606339455
0.0004916153848171234
```

and the code for this: 

```
import timeit
class SymbolOld:
    def __new__(self, name):
        self.name = name

    def __hash__(self):
        return hash(tuple(("Symbol", self.name)))


class SymbolNew:
    def __new__(self, name):
        self.name = name
        self.hashval = hash(tuple(("Symbol", self.name)))

    def __hash__(self):
        return self.hashval

old = SymbolOld("old")
new = SymbolOld("new")

d = {old: 1, new: 2, "string": 3}

def test(key): d[key]

ITERATIONS=10000
print(timeit.timeit("test(old)", setup="from __main__ import test, old", number=ITERATIONS))
print(timeit.timeit("test(new)", setup="from __main__ import test, new", number=ITERATIONS))
print(timeit.timeit("test('string')", setup="from __main__ import test", number=ITERATIONS))
```